Complete migrations to ecs, add few tests
