+++ src/core/systems.py
@@ -17,10 +17,12 @@ from core.components import (
     UnderUserControl,
     AI,
 )
+from core.board import (
+    get_fov_mask,
+)
 
 
 BOARD_SIZE = 10
-Distance = NewType('Distance', int)
 
 
 @dataclass(frozen=True)
@@ -200,75 +202,19 @@ class CleanupSystem(System):
                 )
 
 
+class PositionWrapper:
+    def __init__(self, position: Position):
+        self._position = position
+
+
 class FOVSystem(System):
     def process(self):
         for entity in self._manager.entities.filter([Vision, Position]):
             position = self._manager.entities.get(entity, Position)
-            fov = self._get_fov_mask(position)
+            is_full = lambda x: self._manager.components.get(x)
+            fov = get_fov_mask(position=position, fow_size=3, board_size=BOARD_SIZE, is_full=is_full)
             yield Event(
                 entity=entity,
                 component_class=FOV,
                 func=SetFOV(fov),
             )
-
-    def _generate_movements(self, pos: Position, distance: Distance) -> Iterable[Position]:
-        for mx in range(-distance, distance + 1):
-            for my in range(-distance, distance + 1):
-                new_pos = Position(mx + pos.x, my + pos.y)
-                yield new_pos
-
-    def _get_fov_mask(self, position: Position):
-        def is_visible(pos_to: Position) -> bool:
-            line = list(self._get_line_of_view(position, pos_to))
-            for i, pos_for_check in enumerate(line):
-                if i not in (0, len(line) - 1) and self._manager.components.get(pos_for_check):
-                    return False
-            return True
-
-        positions = self._generate_movements(position, distance=Distance(3))
-        result = set()
-        for pos in positions:
-            if not (0 <= pos.x < BOARD_SIZE and 0 <= pos.y < BOARD_SIZE):
-                continue
-            if not is_visible(pos):
-                continue
-            result.add(pos)
-        return result
-
-    def _get_line_of_view(self, pos1: Position, pos2: Position) -> Iterable[Position]:
-        """
-        Bresenham's line algorithm
-
-        There may be some problems, check it again and add tests
-        """
-
-        # TODO: check it again and add tests
-        delta_x = abs(pos2.x - pos1.x)
-        delta_y = abs(pos2.y - pos1.y)
-        if delta_x > delta_y:
-            a1, b1, a2, b2 = pos1.x, pos1.y, pos2.x, pos2.y
-        else:
-            a1, b1, a2, b2 = pos1.y, pos1.x, pos2.y, pos2.x
-        delta_a = abs(a2 - a1)
-        delta_b = abs(b2 - b1)
-        error = 0.0
-        delta_err = delta_b / delta_a if delta_a != 0 else 0
-        b = b1
-        direction = b2 - b1
-        if direction > 0:
-            direction = 1
-        if direction < 0:
-            direction = -1
-        if a1 < a2:
-            range_a = range(a1, a2 + 1)
-        else:
-            range_a = range(a2, a1 + 1)[::-1]
-        for a in range_a:
-            if delta_x > delta_y:
-                yield Position(a, b)
-            else:
-                yield Position(b, a)
-            error = error + delta_err
-            if error >= 0.5:
-                b = b + direction
-                error = error - 1.0
